Modified GENN-A * solver
